other projects/OccilanStats-6-main/OccilanStats-6-main/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2